Revise how dice sets are defined for `sum` command
Reworked the splitting and construction of DiceSets to simplify
the management of the sets and bundling bonuses with the DiceSet
nearest each bonus to the left instead of tracking them as
separate "rolls" in their own DiceSet.

Output was updated to only show rolled dice (excluding bonuses)
and the total.